requirements.txt
flask
flask-cors
google-generativeai
pinecone
gunicorn